audio_service/requirements.txt
pyaudio
openai-whisper
paho-mqtt
